pump_dump_monitor/src/requirements.txt
requests==2.32.3
numpy==1.26.4
python-dotenv==0.21.0
schedule==1.2.2